[refactor] #344 LoginService V2 명칭 변경
* 리팩토링 완료로 인해 V2 -> V1으로 변경
