+++ app/src/main/java/com/runnect/runnect/data/service/LoginService.kt
@@ -1,11 +1,15 @@
 package com.runnect.runnect.data.service
 
-import com.runnect.runnect.data.dto.response.ResponseGetRefreshToken
-import retrofit2.http.*
+import com.runnect.runnect.data.dto.request.RequestPostLogin
+import com.runnect.runnect.data.dto.response.ResponsePostLogin
+import retrofit2.http.Body
+import retrofit2.http.POST
 
 interface LoginService {
 
-    //토큰 재발급
-    @GET("/api/auth/getNewToken")
-    suspend fun getNewToken(): ResponseGetRefreshToken
+    //로그인
+    @POST("/api/auth")
+    suspend fun postLogin(
+        @Body request: RequestPostLogin
+    ): Result<ResponsePostLogin>
 }
