+++ app/src/main/java/com/runnect/runnect/data/service/TokenService.kt
@@ -1,0 +1,11 @@
+package com.runnect.runnect.data.service
+
+import com.runnect.runnect.data.dto.response.ResponseGetRefreshToken
+import retrofit2.http.*
+
+interface TokenService {
+
+    //토큰 재발급
+    @GET("/api/auth/getNewToken")
+    suspend fun getNewToken(): ResponseGetRefreshToken
+}
