+++ app/src/main/java/com/runnect/runnect/data/source/remote/RemoteLoginDataSource.kt
@@ -2,13 +2,13 @@ package com.runnect.runnect.data.source.remote
 
 import com.runnect.runnect.data.dto.request.RequestPostLogin
 import com.runnect.runnect.data.dto.response.ResponsePostLogin
-import com.runnect.runnect.data.service.v2.LoginV2Service
+import com.runnect.runnect.data.service.LoginService
 import javax.inject.Inject
 
 class RemoteLoginDataSource @Inject constructor(
-    private val loginV2Service: LoginV2Service
+    private val loginService: LoginService
 ) {
     suspend fun postLogin(
         requestPostLogin: RequestPostLogin
-    ): Result<ResponsePostLogin> = loginV2Service.postLogin(requestPostLogin)
+    ): Result<ResponsePostLogin> = loginService.postLogin(requestPostLogin)
 }
